# [Storage] Optimize storage.AccountCoin Encoding