This PR changes `CoinStorage` to use an optimized encoding scheme instead of the standard `zstd` encoding. This significantly reduces the size of data stored on disk, utilizes much less memory, and is waaaaaay faster.

### Size Comparison (81% reduction in size 🎉 )
```text
Small Data (happy path):
Uncompressed: 147 B, Standard Compressed: 116 B, Optimized: 21 B

Complex Data (tons of metadata):
Uncompressed: 207 B, Standard Compressed: 137 B, Optimized: 41 B
```

### Runtime Benchmarks (99.5% reduction in time/op 🎉 )
```text
BenchmarkAccountCoinStandard
BenchmarkAccountCoinStandard-12            	   15799	     74628 ns/op	    2793 B/op	      29 allocs/op
BenchmarkComplexAccountCoinStandard
BenchmarkComplexAccountCoinStandard-12     	   15390	     80565 ns/op	    3597 B/op	      44 allocs/op
BenchmarkAccountCoinOptimized
BenchmarkAccountCoinOptimized-12           	 3039498	       373 ns/op	     192 B/op	      10 allocs/op
BenchmarkComplexAccountCoinOptimized
BenchmarkComplexAccountCoinOptimized-12    	  572340	      2071 ns/op	    1696 B/op	      34 allocs/op
```